Add files via upload
I recommend using the celeba dataset or another larger dataset. Also that it is racially diverse because the dataset I used was the gender classification dataset on kaggle. I dont recommend it unless you are just trying to save storage room due to its smaller size. My model predicted wrong gender on some other races of woman that were not well represented in this dataset as they might be in others. Just a precaution. Other than that very high training and validation accuracy as well as low test and validation loss. Most of the time predicted 100% correct. The choice is up to you which daset you would like to work with.